refactor: rename functions -> utils
