+++ src/test/jsonUseCase.test.ts
@@ -7,6 +7,6 @@ import { Entry } from '../entryEntity';
 import * as E from 'fp-ts/lib/Either';
 import * as TE from 'fp-ts/lib/TaskEither';
 import { pipe } from 'fp-ts/lib/function';
-import { tap } from '../functions';
+import { tap } from '../utils';
 
 describe('jsonUseCase', () => {});
